Better safe area handling everywhere
